+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - `FtpHandler` 增加新方法。
   - FtpHandler.clearDirectory。
   - FtpHandler.copyFile。
+  - FtpHandler.descFile。
 
 - 优化部分类的文档注释。
   - com.dwarfeng.ftp.handler.FtpHandler。